backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
mysql-connector-python @ http://cdn.mysql.com/Downloads/Connector-Python/mysql-connector-python-2.0.4.zip
Pillow==9.1.1
PyJWT==2.4.0
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
django-cors-headers==3.13.0
gunicorn==20.1.0
